Docs: Main page framework and comments
Created main, plain structure for the page. Created alt text, but did not add images yet as I need to take relevant photos this weekend.
Still need to add footer, but should push this so far in order to have a version to revert to if anything funky happens with the code.